+++ driver/web/docs/gen/gen_types.go
@@ -106,18 +106,25 @@ const (
 	ServicesReqCredentialSendVerifyAuthTypeEmail ServicesReqCredentialSendVerifyAuthType = "email"
 )
 
-// Defines values for ServicesReqServiceAccountsParamsRequestAuthType.
+// Defines values for ServicesReqServiceAccountsAccessTokenAuthType.
 const (
-	ServicesReqServiceAccountsParamsRequestAuthTypeSignature ServicesReqServiceAccountsParamsRequestAuthType = "signature"
+	ServicesReqServiceAccountsAccessTokenAuthTypeSignature ServicesReqServiceAccountsAccessTokenAuthType = "signature"
 
-	ServicesReqServiceAccountsParamsRequestAuthTypeStaticToken ServicesReqServiceAccountsParamsRequestAuthType = "static_token"
+	ServicesReqServiceAccountsAccessTokenAuthTypeStaticToken ServicesReqServiceAccountsAccessTokenAuthType = "static_token"
 )
 
-// Defines values for ServicesReqServiceAccountsTokenRequestAuthType.
+// Defines values for ServicesReqServiceAccountsAccessTokensAuthType.
 const (
-	ServicesReqServiceAccountsTokenRequestAuthTypeSignature ServicesReqServiceAccountsTokenRequestAuthType = "signature"
+	ServicesReqServiceAccountsAccessTokensAuthTypeSignature ServicesReqServiceAccountsAccessTokensAuthType = "signature"
 
-	ServicesReqServiceAccountsTokenRequestAuthTypeStaticToken ServicesReqServiceAccountsTokenRequestAuthType = "static_token"
+	ServicesReqServiceAccountsAccessTokensAuthTypeStaticToken ServicesReqServiceAccountsAccessTokensAuthType = "static_token"
+)
+
+// Defines values for ServicesReqServiceAccountsParamsAuthType.
+const (
+	ServicesReqServiceAccountsParamsAuthTypeSignature ServicesReqServiceAccountsParamsAuthType = "signature"
+
+	ServicesReqServiceAccountsParamsAuthTypeStaticToken ServicesReqServiceAccountsParamsAuthType = "static_token"
 )
 
 // Defines values for ServicesResAuthorizeServiceTokenType.
@@ -769,35 +776,42 @@ type ServicesReqCredentialUpdate struct {
 	Params            *interface{} `json:"params,omitempty"`
 }
 
-// Service account token for auth_type="static_token"
-type ServicesReqServiceAccountsCredsStaticToken struct {
-	Token string `json:"token"`
-}
-
-// ServicesReqServiceAccountsParamsRequest defines model for _services_req_service-accounts_ParamsRequest.
-type ServicesReqServiceAccountsParamsRequest struct {
-	AuthType ServicesReqServiceAccountsParamsRequestAuthType `json:"auth_type"`
+// ServicesReqServiceAccountsAccessToken defines model for _services_req_service-accounts_access-token.
+type ServicesReqServiceAccountsAccessToken struct {
+	AccountId string                                        `json:"account_id"`
+	AppId     *string                                       `json:"app_id"`
+	AuthType  ServicesReqServiceAccountsAccessTokenAuthType `json:"auth_type"`
 
 	// Service account token for auth_type="static_token"
-	Creds *ServicesReqServiceAccountsCredsStaticToken `json:"creds,omitempty"`
+	Creds *ServicesServiceAccountsCredsStaticToken `json:"creds,omitempty"`
+	OrgId *string                                  `json:"org_id"`
 }
 
-// ServicesReqServiceAccountsParamsRequestAuthType defines model for ServicesReqServiceAccountsParamsRequest.AuthType.
-type ServicesReqServiceAccountsParamsRequestAuthType string
+// ServicesReqServiceAccountsAccessTokenAuthType defines model for ServicesReqServiceAccountsAccessToken.AuthType.
+type ServicesReqServiceAccountsAccessTokenAuthType string
 
-// ServicesReqServiceAccountsTokenRequest defines model for _services_req_service-accounts_TokenRequest.
-type ServicesReqServiceAccountsTokenRequest struct {
+// ServicesReqServiceAccountsAccessTokens defines model for _services_req_service-accounts_access-tokens.
+type ServicesReqServiceAccountsAccessTokens struct {
 	AccountId string                                         `json:"account_id"`
-	AppId     *string                                        `json:"app_id"`
-	AuthType  ServicesReqServiceAccountsTokenRequestAuthType `json:"auth_type"`
+	AuthType  ServicesReqServiceAccountsAccessTokensAuthType `json:"auth_type"`
 
 	// Service account token for auth_type="static_token"
-	Creds *ServicesReqServiceAccountsCredsStaticToken `json:"creds,omitempty"`
-	OrgId *string                                     `json:"org_id"`
+	Creds *ServicesServiceAccountsCredsStaticToken `json:"creds,omitempty"`
 }
 
-// ServicesReqServiceAccountsTokenRequestAuthType defines model for ServicesReqServiceAccountsTokenRequest.AuthType.
-type ServicesReqServiceAccountsTokenRequestAuthType string
+// ServicesReqServiceAccountsAccessTokensAuthType defines model for ServicesReqServiceAccountsAccessTokens.AuthType.
+type ServicesReqServiceAccountsAccessTokensAuthType string
+
+// ServicesReqServiceAccountsParams defines model for _services_req_service-accounts_params.
+type ServicesReqServiceAccountsParams struct {
+	AuthType ServicesReqServiceAccountsParamsAuthType `json:"auth_type"`
+
+	// Service account token for auth_type="static_token"
+	Creds *ServicesServiceAccountsCredsStaticToken `json:"creds,omitempty"`
+}
+
+// ServicesReqServiceAccountsParamsAuthType defines model for ServicesReqServiceAccountsParams.AuthType.
+type ServicesReqServiceAccountsParamsAuthType string
 
 // ServicesResAccountAuthTypeLink defines model for _services_res_account_auth-type-link.
 type ServicesResAccountAuthTypeLink struct {
@@ -819,6 +833,18 @@ type ServicesResAuthorizeService struct {
 
 // The type of the provided tokens to be specified when they are sent in the "Authorization" header
 type ServicesResAuthorizeServiceTokenType string
+
+// ServicesResServiceAccountsAccessTokens defines model for _services_res_service-accounts_access-tokens.
+type ServicesResServiceAccountsAccessTokens struct {
+	AppId *string               `json:"app_id"`
+	OrgId *string               `json:"org_id"`
+	Token SharedResRokwireToken `json:"token"`
+}
+
+// Service account token for auth_type="static_token"
+type ServicesServiceAccountsCredsStaticToken struct {
+	Token string `json:"token"`
+}
 
 // SharedReqAccountCheck defines model for _shared_req_AccountCheck.
 type SharedReqAccountCheck struct {
@@ -1310,10 +1336,13 @@ type PostAdminAuthRefreshJSONBody SharedReqRefresh
 type PostAdminAuthVerifyMfaJSONBody SharedReqMfa
 
 // PostBbsAccessTokenJSONBody defines parameters for PostBbsAccessToken.
-type PostBbsAccessTokenJSONBody ServicesReqServiceAccountsTokenRequest
+type PostBbsAccessTokenJSONBody ServicesReqServiceAccountsAccessToken
+
+// PostBbsAccessTokensJSONBody defines parameters for PostBbsAccessTokens.
+type PostBbsAccessTokensJSONBody ServicesReqServiceAccountsAccessTokens
 
 // PostBbsServiceAccountIdJSONBody defines parameters for PostBbsServiceAccountId.
-type PostBbsServiceAccountIdJSONBody ServicesReqServiceAccountsParamsRequest
+type PostBbsServiceAccountIdJSONBody ServicesReqServiceAccountsParams
 
 // GetBbsServiceRegsParams defines parameters for GetBbsServiceRegs.
 type GetBbsServiceRegsParams struct {
@@ -1638,10 +1667,13 @@ type PostSystemServiceRegsJSONBody ServiceReg
 type PutSystemServiceRegsJSONBody ServiceReg
 
 // PostTpsAccessTokenJSONBody defines parameters for PostTpsAccessToken.
-type PostTpsAccessTokenJSONBody ServicesReqServiceAccountsTokenRequest
+type PostTpsAccessTokenJSONBody ServicesReqServiceAccountsAccessToken
+
+// PostTpsAccessTokensJSONBody defines parameters for PostTpsAccessTokens.
+type PostTpsAccessTokensJSONBody ServicesReqServiceAccountsAccessTokens
 
 // PostTpsServiceAccountIdJSONBody defines parameters for PostTpsServiceAccountId.
-type PostTpsServiceAccountIdJSONBody ServicesReqServiceAccountsParamsRequest
+type PostTpsServiceAccountIdJSONBody ServicesReqServiceAccountsParams
 
 // GetTpsServiceRegsParams defines parameters for GetTpsServiceRegs.
 type GetTpsServiceRegsParams struct {
@@ -1717,6 +1749,9 @@ type PostAdminAuthVerifyMfaJSONRequestBody PostAdminAuthVerifyMfaJSONBody
 
 // PostBbsAccessTokenJSONRequestBody defines body for PostBbsAccessToken for application/json ContentType.
 type PostBbsAccessTokenJSONRequestBody PostBbsAccessTokenJSONBody
+
+// PostBbsAccessTokensJSONRequestBody defines body for PostBbsAccessTokens for application/json ContentType.
+type PostBbsAccessTokensJSONRequestBody PostBbsAccessTokensJSONBody
 
 // PostBbsServiceAccountIdJSONRequestBody defines body for PostBbsServiceAccountId for application/json ContentType.
 type PostBbsServiceAccountIdJSONRequestBody PostBbsServiceAccountIdJSONBody
@@ -1876,6 +1911,9 @@ type PutSystemServiceRegsJSONRequestBody PutSystemServiceRegsJSONBody
 
 // PostTpsAccessTokenJSONRequestBody defines body for PostTpsAccessToken for application/json ContentType.
 type PostTpsAccessTokenJSONRequestBody PostTpsAccessTokenJSONBody
+
+// PostTpsAccessTokensJSONRequestBody defines body for PostTpsAccessTokens for application/json ContentType.
+type PostTpsAccessTokensJSONRequestBody PostTpsAccessTokensJSONBody
 
 // PostTpsServiceAccountIdJSONRequestBody defines body for PostTpsServiceAccountId for application/json ContentType.
 type PostTpsServiceAccountIdJSONRequestBody PostTpsServiceAccountIdJSONBody